Get Latest

+++ Solutions/PowerShell.Commands/Commands/Fields/RemoveField.cs
@@ -19,33 +19,40 @@ namespace OfficeDevPnP.PowerShell.Commands
 
         protected override void ExecuteCmdlet()
         {
-            var list = this.SelectedWeb.GetList(List);
-
-            Field f = Identity.Field;
-            if (list != null)
+            if (List != null)
             {
-                if (f == null)
+                var list = this.SelectedWeb.GetList(List);
+
+                Field f = Identity.Field;
+                if (list != null)
                 {
-                    if (Identity.Id != Guid.Empty)
+                    if (f == null)
                     {
-                        f = list.Fields.GetById(Identity.Id);
+                        if (Identity.Id != Guid.Empty)
+                        {
+                            f = list.Fields.GetById(Identity.Id);
+                        }
+                        else if (!string.IsNullOrEmpty(Identity.Name))
+                        {
+                            f = list.Fields.GetByInternalNameOrTitle(Identity.Name);
+                        }
                     }
-                    else if (!string.IsNullOrEmpty(Identity.Name))
+                    ClientContext.Load(f);
+                    ClientContext.ExecuteQuery();
+                    if (f != null && f.IsPropertyAvailable("InternalName"))
                     {
-                        f = list.Fields.GetByInternalNameOrTitle(Identity.Name);
+                        if (Force || ShouldContinue(string.Format(Properties.Resources.DeleteField0, f.InternalName), Properties.Resources.Confirm))
+                        {
+                            f.DeleteObject();
+                            ClientContext.ExecuteQuery();
+                        }
                     }
                 }
-                if (f != null)
-                {
-                    if (Force || ShouldContinue(string.Format(Properties.Resources.DeleteField0, f.InternalName), Properties.Resources.Confirm))
-                    {
-                        f.DeleteObject();
-                        ClientContext.ExecuteQuery();
-                    }
-                }
-            }
+            } 
             else
             {
+                Field f = Identity.Field;
+
                 if (f == null)
                 {
                     if (Identity.Id != Guid.Empty)
@@ -57,7 +64,10 @@ namespace OfficeDevPnP.PowerShell.Commands
                         f = this.SelectedWeb.Fields.GetByInternalNameOrTitle(Identity.Name);
                     }
                 }
-                if(f != null)
+                ClientContext.Load(f);
+                ClientContext.ExecuteQuery();
+
+                if (f != null && f.IsPropertyAvailable("InternalName"))
                 {
                     if (Force || ShouldContinue(string.Format(Properties.Resources.DeleteField0, f.InternalName), Properties.Resources.Confirm))
                     {
